remove types file
this now uses the structure that the schedules action/reducer has (other PR)
